人机协同/digital-employee-closed-loop/pyproject.toml
[project]
name = "digital-employee-closed-loop"
version = "0.1.0"
description = "Closed-loop skills for verify/qc digital employees and human operation feedback"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "pydantic>=2.7.0",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.pytest.ini_options]
python_files = ["test_*.py"]
addopts = "-q"
